#328: Moved the consumable models to model/items/consumable
-
