+++ src/app/ffbe/mappers/effects/effect-parser.ts
@@ -3,7 +3,7 @@ import {FFBE_CATEGORIES_OBJETS, FFBE_ELEMENTS, FFBE_FRENCH_TABLE_INDEX} from '..
 import {CategorieObjet} from '../../model/objet/categorie-objet.model';
 import {Esper} from '../../model/esper.model';
 import {Equipment} from '../../model/items/equipment/equipment.model';
-import {Consumable} from '../../model/consumable.model';
+import {Consumable} from '../../model/items/consumable/consumable.model';
 import {Element} from '../../model/element.model';
 import {TargetPrepositionEnum} from '../../model/effects/target-preposition.enum';
 import {SkillEffect} from '../../model/effects/skill-effect.model';
