+++ src/app/ffbe/model/items/consumable/consumable-strings.model.ts
