+++ src/app/ffbe/model/items/consumable/consumable.model.ts
