+++ src/app/ffbe/services/consumables.service.ts
@@ -1,6 +1,6 @@
 import {DataMiningClientService} from './data-mining-client.service';
 import {Injectable} from '@angular/core';
-import {Consumable} from '../model/consumable.model';
+import {Consumable} from '../model/items/consumable/consumable.model';
 import {FFBE_FRENCH_TABLE_INDEX} from '../ffbe.constants';
 
 @Injectable()
